% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop.R
\docType{data}
\name{crop}
\alias{crop}
\title{Crop data containing hypothetical sowing flowering and harvest dates}
\format{
A data frame with 10 rows and 3 variables:
\describe{
\item{sowing_date}{Sowing date of crop}
\item{flower_date}{Flowering date of crop}
\item{harvest_date}{Harvest date of crop}
}
}
\source{
Hypothetical sowing, flowering and harvest dates based on typical horticultural crops in Queensland
}
\usage{
crop
}
\description{
A dataset containing typical dates for agronomic data
}
\details{
Typical dates for crop data sets that would typically contain
agronomic traits such as yield, dry matter yield and so
on. However, here the dataset only contains dates which are used
to demonstrate how to add computed weather variables either
starting or ending at a specified date or between two dates.
}
\keyword{datasets}
