% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_of_year.R
\name{number_of_days}
\alias{number_of_days}
\title{The number of days between two dates}
\usage{
number_of_days(x, start)
}
\arguments{
\item{x}{The end date with class \code{\link{Date}}}

\item{start}{The start date with class \code{\link{Date}}}
}
\value{
A \code{\link[base]{numeric}} variable containing the number
of days between the two dates
}
\description{
A convenience function which is simply \code{as.numeric(end_date -
start_date)}
}
\examples{
library(lubridate)
number_of_days(x = ymd("2021-01-05"), start = ymd("2020-12-28"))
}
